Move away from deprecated GTimeVal structs
These were marked as deprecated in GLib 2.62. Their replacements have
been in GLib since 2.26, which was released on 2010, so we should be
fine to just drop the deprecated funcs entirely instead of trying to
maintain compatibility with the old version.

Also update debian/compat to allow building the pacakge under bullseye
and clean up a few related lintian warnings.

Fixes #27